Break out components into files
