+++ src/js/personalization/account-beta/components/LoginSettings.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function LoginSettings() {
+  return (
+    <div>
+      <h4>Login Settings</h4>
+      <p>Want to change your email, password, or other account settings?<br/>
+        <a href="https://wallet.id.me/settings" target="_blank">Go to ID.me to manage your account</a>
+      </p>
+    </div>
+  );
+}
