+++ src/js/personalization/account-beta/components/UserDataSection.jsx
@@ -3,7 +3,6 @@ import { connect } from 'react-redux';
 import AcceptTermsPrompt from '../../../common/components/AcceptTermsPrompt';
 import LoadingIndicator from '../../../common/components/LoadingIndicator';
 import Modal from '../../../common/components/Modal';
-import AlertBox from '../../../common/components/AlertBox';
 
 import { getMultifactorUrl, handleMultifactor } from '../../../common/helpers/login-helpers';
 import { updateMultifactorUrl } from '../../../login/actions';
@@ -13,78 +12,10 @@ import {
   acceptTerms,
 } from '../actions';
 
-function MultifactorMessage({ multifactor, handleMultifactorRequest }) {
-  const headline = 'Add extra security to your account';
-  const content = (
-    <div>
-      <p>For additional protection, we encourage you to add a second security step for signing in to your account.</p>
-      <button className="usa-button usa-button-secondary" onClick={handleMultifactorRequest}>Add security step</button>
-    </div>
-  );
-
-  return (
-    <AlertBox
-      headline={headline}
-      content={content}
-      isVisible={!multifactor}
-      status="info"/>
-  );
-}
-
-function AccountVerification({ accountType }) {
-  let content = null;
-
-  if (accountType !== 3) {
-    content = (
-      <div>
-        <p>Verify your identity to access more services your may be eligible for.<br/>
-          <a className="usa-button-primary" href="/verify?next=/profile">Verify identity</a>
-        </p>
-      </div>
-    );
-  } else {
-    content = <p><i className="fa fa-check-circle"/> Your account has been verified.</p>;
-  }
-
-  return (
-    <div>
-      <h4>Account verification</h4>
-      {content}
-    </div>
-  );
-}
-
-function LoginSettings() {
-  return (
-    <div>
-      <h4>Login Settings</h4>
-      <p>Want to change your email, password, or other account settings?<br/>
-        <a href="https://wallet.id.me/settings" target="_blank">Go to ID.me to manage your account</a>
-      </p>
-    </div>
-  );
-}
-
-function TermsAndConditions({ healthTermsCurrent, openModal }) {
-  let content = null;
-
-  if (healthTermsCurrent) {
-    content = (
-      <p>You have accepted the latest health terms and conditions for this site.</p>
-    );
-  } else {
-    content = (
-      <p>In order to refill your prescriptions, message your health care team, and get your VA health records, you need to accept the <a onClick={openModal}>Terms and Conditions for Health Tools</a>.</p>
-    );
-  }
-
-  return (
-    <div>
-      <h4>Terms and conditions</h4>
-      {content}
-    </div>
-  );
-}
+import AccountVerification from './AccountVerification';
+import LoginSettings from './LoginSettings';
+import MultifactorMessage from './MultifactorMessage';
+import TermsAndConditions from './TermsAndConditions';
 
 class UserDataSection extends React.Component {
   constructor(props) {
